feat: support for custom language service plugins
