+++ packages/vue-language-service/src/index.ts
@@ -1,8 +1,9 @@
+export * from '@volar/vue-language-service-types';
 export * from '@volar/vue-typescript';
 export * from './documentService';
-export * from './languageService';
-export { SemanticToken } from './utils/definePlugin';
-export { margeWorkspaceEdits } from './languageFuatures/rename';
 export { executePluginCommand, ExecutePluginCommandArgs } from './languageFuatures/executeCommand';
-export { convertTagNameCasingCommand, ConvertTagNameCasingCommandArgs } from './vuePlugins/tagNameCasingConversions';
+export { margeWorkspaceEdits } from './languageFuatures/rename';
+export * from './languageService';
 export * from './types';
+export { convertTagNameCasingCommand, ConvertTagNameCasingCommandArgs } from './vuePlugins/tagNameCasingConversions';
+
