+++ packages/vue-language-service/src/vuePlugins/autoWrapParentheses.ts
@@ -1,6 +1,6 @@
 import type { TextDocument } from 'vscode-languageserver-textdocument';
 import * as vscode from 'vscode-languageserver-protocol';
-import { EmbeddedLanguagePlugin } from '../utils/definePlugin';
+import { EmbeddedLanguagePlugin } from '@volar/vue-language-service-types';
 import { VueDocument } from '@volar/vue-typescript';
 import { isCharacterTyping } from './autoCompleteRefs';
 import * as shared from '@volar/shared';
